Add defensive scripting tests and replace QtConcurrent runs with QThreadPool.Harden script sandbox and expand coverage
Replace QtConcurrent calls with QThreadPool, fix a signed-size
comparison, and add tests for sandboxing, timeouts, and script limits.

+++ src/app/KWinCursorTracker.cpp
@@ -97,7 +97,7 @@ void KWinCursorTracker::queryGlobal(const std::function<void(const QPoint &)> &c
         callback(QCursor::pos());
         return;
     }
-    if (script.write(kCursorScript) != qstrlen(kCursorScript) || !script.flush()) {
+    if (script.write(kCursorScript) != qint64(qstrlen(kCursorScript)) || !script.flush()) {
         tracker->m_kwinUnavailable = true;
         callback(QCursor::pos());
         return;
